Changed mpfr_sum to use unsigned long instead of unsigned int. More tests.
[From SVN r2775 (trunk)]

+++ mpfr-impl.h
@@ -518,7 +518,7 @@ int mpfr_round_raw_4 _MPFR_PROTO ((mp_limb_t *, mp_limb_t *,
 int mpfr_check _MPFR_PROTO ((mpfr_srcptr));
 int mpfr_sub1sp _MPFR_PROTO ((mpfr_ptr, mpfr_srcptr, mpfr_srcptr, mp_rnd_t));
 
-void mpfr_count_sort _MPFR_PROTO ((mpfr_ptr const tab[], unsigned int n,
+void mpfr_count_sort _MPFR_PROTO ((mpfr_ptr const tab[], unsigned long n,
 				   mpfr_srcptr *perm));
 
 #if defined (__cplusplus)
